v1.8
-added config option: Set Respawn Point
  -if set, you can right-click campfires to set your spawn
  -campfires must be lit to act as a respawn point
-added config option: Burn Out on Respawn
  -if set, campfires will burn out when you use them to respawn
-this feature is from [Home Camp](https://www.curseforge.com/minecraft/mc-mods/home-camp)

-added galacticraft & advanced rocketry compatibility
  -campfires will burn out if they don't have oxygen
  -campfire smoke will rise at different rates depending on the atmosphere

-comparators now detect the fullness of a campfire's inventory

-campfires can now be rotated by tools that rotate blocks
  -holding sneak will show the direction a campfire is facing in WAILA

-/campfirebackport has another subcommand: locationinfo
  -prints the biome and dimension IDs at the player or at coordinates x z

-bug fixes
  -burn out timers / regen timers set to start at 0 won't get stuck anymore
  -custom recipe inputs based on Java Classes that represent Blocks work again

-technical
  -when a campfire is going to be extinguished/ignited, an event is posted to the forge event bus. you can cancel the event to stop the campfire from being extinguished/ignited.

+++ src/main/java/connor135246/campfirebackport/client/rendering/RenderItemBlockCampfire.java
@@ -42,10 +42,9 @@ public class RenderItemBlockCampfire implements IItemRenderer
         if (tilenbt.hasKey(TileEntityCampfire.KEY_Life) && tilenbt.hasKey(TileEntityCampfire.KEY_StartingLife)
                 && tilenbt.hasKey(TileEntityCampfire.KEY_PreviousTimestamp))
         {
-            float life = (float) tilenbt.getInteger(TileEntityCampfire.KEY_Life);
-            float timePassed = (float) (Minecraft.getMinecraft().thePlayer.worldObj.getTotalWorldTime()
-                    - tilenbt.getLong(TileEntityCampfire.KEY_PreviousTimestamp));
-            float starting = (float) tilenbt.getInteger(TileEntityCampfire.KEY_StartingLife);
+            float life = tilenbt.getInteger(TileEntityCampfire.KEY_Life);
+            float timePassed = Minecraft.getMinecraft().thePlayer.worldObj.getTotalWorldTime() - tilenbt.getLong(TileEntityCampfire.KEY_PreviousTimestamp);
+            float starting = tilenbt.getInteger(TileEntityCampfire.KEY_StartingLife);
 
             size = MathHelper.clamp_float((life - timePassed) / starting, 0.01F, 1.0F);
         }
